This is new
The new code starts off colluding and when we reach the 100th turn it
returns betray, if their last 2 turns are betray then we return betray,
if their last 3 plays are collude we return collude. On the last turn we
return betray, and if none of those statement come back true then we
default return collude.