We try to get in the head of our opponents.
We will not change our strategy because we got 2nd place. Only the P6
guy beat us. Our strategy was to beat the guy on his terms, and keep as
many points as possible. We need to change nothing.
Overall our coding was pretty good. We don't need to change much
because the only problems we had was with the collude unless betrayed
player. We did add an if they collude twice, we collude so we can beat
the collude unless betrayed guy.
The results went really good for us. We got 3rd place and we weren't
that far from being higher up. We don't believe that we need to change
anything because we did good in the results.